Start on Gc code refactor
